+++ resources/js/components/RoleTable.vue
@@ -25,7 +25,7 @@ import {
   RotateCcwIcon
 } from 'lucide-vue-next'
 import LocaleBadge from './LocaleBadge.vue'
-import type { Role } from '@/types'
+import type { Role } from '@/laravel-roles/types'
 
 defineProps<{
   roles: Role[]